songcheat-core 1.0.21
Fixed the issues arising when lyrics and units do not match exactly, i.e. when either:
A. Lyrics for a unit actually start at the end of previous unit
B. Lyrics for a unit actually end at the start of next unit

One can now start the unit lyrics with either :
A. ">" : means that the last lyrics line of previous unit will be prepended to this unit
B. "<": means that the first lyrics line of this unit will be appended to previous unit

Compiler: method compileUnit handles the ">" resp. "<" found at the start of a unit, converted into a rightArrow resp. leftArrow property on the group.

Ascii: method getUnitText renamed as getUnitsText which takes an array of several units and returns an array of corresponding texts.
Working on the whole array allows handling the left/rightArrow properties possibly set on the first group of each unit to detect where we should actually split unit texts.
Chord changes and bar signs are still colored using the color of the unit they really belong to (which means we can have two colors on a same line).

Fully compatible with the existing: if source does not change, result will be exactly the same as before (proof: .ascii files in songcheat-demos have not changed).

This makes the following workarounds obsolete:
1/ Paroles entre parenthèses au début de l'unité pour signifier qu'elles appartiennent à l'unité précédente
2/ Concaténer les unités afin de contourner ce problème
3/ Adapter les paroles aux unités (déplacer le début/fin de chaque couplet à la fin/au début de l'unité précédente/suivante)
4/ Adapter les unités aux paroles (déplacer la dernière/première mesure d'une unité à la suivante/précédente)

Removed those workarounds in the following songcheats:
1/A. Losing my Religion
2/B. Mon ami l'animal
1/A. Droit dans le soleil
1/A. Jardin d'hiver
1/A. L'Hippopodame (anciennement A3.)
4/A. La Brise
2/B. Tant de nuits
2/A. Balade à Toronto